requirements.txt
# Django Core
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0

# Database
psycopg2-binary==2.9.9

# CORS & Security
django-cors-headers==4.3.1

# Rate Limiting
django-ratelimit==4.1.0
redis==5.0.1
django-redis==5.4.0

# Environment
python-dotenv==1.0.0

# Development Tools
django-extensions==3.2.3
django-debug-toolbar==4.2.0
ipython==8.17.2

# HTTP Client for n8n integration
requests==2.31.0

# Password Hashing (Argon2)
argon2-cffi==23.1.0

# ASGI Support
uvicorn[standard]==0.24.0

